requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.23.2
requests==2.31.0
APScheduler==3.11.0
python-dotenv==1.0.0
